Request of inactive user
