+++ app/index.html
@@ -75,6 +75,7 @@
   <!-- User -->
   <script src="user/userService.js"></script>
   <script src="user/userInactiveController.js"></script>
+  <script src="user/requestInvitationController.js"></script>
 
   <!-- Error Library -->
   <script src="error/errorController.js"></script>
