+++ app/user/request_invitation_dialog.html
@@ -23,7 +23,7 @@
 			    </div>
 		    </md-list-item>
 		    <md-list-item ng-if="requestInvCtrl.showMessage()" ng-click="null">
-		    	<div class="md-list-item-text" layout-align="center center" layout="column" 
+		    	<div class="md-list-item-text" layout-align="center center" layout="column"
 			    style='text-align: center'>
 		            <p>Infelizmente sua instituição ainda não está cadastrada.</p>
 			    </div>
